dissertation_system/requirements.txt
dj_database_url==1.2.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
Django==4.1.5
django_select2==8.1.0
Pillow==9.4.0
pytz==2022.2.1
django_crispy_forms==1.14.0
verify_email==2.4.3